% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalculateStatus.R
\name{recalculateStatus}
\alias{recalculateStatus}
\title{Recalculate the Monthly Status Based on Updates to the Beneficiary
  Object}
\usage{
recalculateStatus(
  beneficiary,
  tier_one_threshold = 250000,
  tier_two_threshold = 5e+05,
  tier_three_threshold = 1e+06,
  max_monthly_contribution = 625,
  poverty_floor = 15000,
  income_multiplier_cap = 120000
)
}
\arguments{
\item{beneficiary}{A `beneficiary` object.}

\item{tier_one_threshold}{`numeric(1)`. The threshold for receiving 
Tier 1 distributions.}

\item{tier_two_threshold}{`numeric(1)`. The threshold for receiving 
Tier 2 distributions.}

\item{tier_three_threshold}{`numeric(1)`. The threshold for receiving 
Tier 3 distributions.}

\item{max_monthly_contribution}{`numeric(1)`. The maximum monthly 
contribution permitted to a Roth IRA per IRS guidelines. The monthly
distribution is based on this amount.}

\item{poverty_floor}{`numeric(1)`. The poverty threshold marking the minimum
for annual Tier 3 distributions.}
}
\description{
Recalculates the Monthly Status and returns the new 
  beneficiary object.
}
