% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBeneficiaryIncome.R
\name{setBeneficiaryIncome}
\alias{setBeneficiaryIncome}
\title{Set Beneficiary Annual Income}
\usage{
setBeneficiaryIncome(beneficiary, ...)
}
\arguments{
\item{beneficiary}{An object of class `beneficiary`}

\item{...}{named pairings where the name is the year an income level starts. 
income will be propagated to future years until the next year in which 
income is given. use `2030 = 25000, 2035 = 30000` to set the income at 
$25,000 for 2030, 2031, 2032, 2033, and 2034. Income will be $30,000 for
2035 and all years after.}
}
\description{
Functions for setting annual earned income for a beneficiary
  within the beneficiary object.
}
