% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setIndividualContribution.R
\name{setIndividualContribution}
\alias{setIndividualContribution}
\title{Set Individual Contributions for a Date Range}
\usage{
setIndividualContribution(beneficiary, contribution, start_date, end_date)
}
\arguments{
\item{beneficiary}{A `beneficiary` object.}

\item{contribution}{`numeric(1)`. The value of the monthly contribution 
for the date range.}

\item{start_date}{`Date(1)`. The start of the date range for which the 
monthly contribution will be applied.}

\item{end_date}{`Date(1)`. The end of the date range for which the montly
contribution will be applied.}
}
\description{
Sets the value of the beneficiary contribution to their 
  personal Roth account for a date range.
}
