% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBeneficiary.R
\name{makeBeneficiary}
\alias{makeBeneficiary}
\title{Make a Beneficiary Object}
\usage{
makeBeneficiary(
  name,
  date_of_birth,
  date_of_entry = date_of_birth + lubridate::years(18)
)
}
\arguments{
\item{name}{`character(1)`. The name of the beneficiary}

\item{date_of_birth}{`Date(1)`. The beneficiary's date of birth.}

\item{date_of_entry}{`Date(1)`. The date on which}
}
\value{
Returns an object of class `beneficiary`. The objects has the
following elements

* `name`: The name of the beneficiary
* `date_of_birth`: The beneficiary's date of birth. 
* `date_of_entry`: The date the beneficiary signs the contract to join the
   trust. 
* `MonthlyStatus`: A `data.frame` giving monthly status of the beneficiary
   account and utiliziation up to age 100.
}
\description{
Creates an object to store projections of wealth fund 
  utilization.
}
